Important revisions to tan:diff()
If a string is considerably long, a search is first done for sequences of unique words that are shared between the strings, then the pieces between are fed to tan:diff(); if the string is not too long, it follows the normal routine, now converted to a two-loop process easier to diagnose. (The older one-loop process has been preserved in this commit, but will likely be deleted in a future commit.) Snap-to-word has been made more reliable.