Clarified the interpretation of special line-end characters.
If a leaf div ends in a special line-end character, the first <tok> or <non-tok>s with @n=1 will
be taken from the next leaf div and fused at the end. If the first element so moved matches
in name the last element of the original leaf div, they will be fused together. That is,
if the leaf div ends with <tok> and the next one starts with a <tok> (or, vice versa, both
<non-tok>) they will become one element. If they are different elements, then the transfer
still happens, but no fusion takes place.