Major changes. See below.
Excerpted from the official guidelines:

The system vocabulary referencing has been streamlined and simplified.
For any element named X with @which, the normalized form of @which points to the normalized form of <name> of some single vocabulary item that has been defined for an element of name X. This is similar to the previous version, except that name normalization now affects _ and - (treated as spacing characters).
For any element named X with an attribute Y that takes idrefs, the space-tokenized values of Y will be checked first against a target @xml:id or @id and, failing that, will check for a match against a <name> in a vocabulary item defined for elements named X.
Implication: <name> is effectively another kind of identifier, like @xml:id or @id. But the latter are case sensitive and matched exactly whereas <name>s are compared after normalization: everything changed to lowercase and hyphens and underscores replaced with spaces.
Implication: attributes like @who may take a mixture of idrefs and names. But if you invoke any names, be sure to replace word spaces with hyphens or underscores, otherwise a single value will be parsed as if multiple ones.
Files are now resolved differently. The goal is to return a file that can be interpreted without needed to reference any vocabularies (including standard TAN ones) or inclusions. (But nothing is done with the <source>s of class 2 files, or other linked files such as <redivision>, <successor>, etc.)
In resolving a file a sharp distinction is made between (1) elements involved in inclusions and (2) those that aren't. All inclusions are resolved recursively and returned to their host including file as a set of <inclusion>s that carry children errors, vocabulary items, and substitutions. All these are applied to the appropriate elements (1) in the including file, and then the rest (2) are resolved. This means that inclusion now brings back not just specific elements, but the vocabulary upon which it depends. It also means that the system of idrefs in included files are now in play. Under the previous system, if you used <resp @include="incl1"> you would have to have also invoked @include for <role> and <person>. It also means that reference errors are now reported in the resolved file directly. Previously these were attended to in terse expansion.