Multiple changes:
Verbs are now treated as exceptional (the same way tokenization patterns are), and promoted from &lt;item> to &lt;verb>. By default, verbs require a subject and an object, and may allow place or period, and do not allow in-lang or at-ref. Users may now specify the construction of a &lt;verb>s through a child &lt;constraint>s, which specify requirements for subject, object, place, period, in-lang, at-ref, each allowed as a child element. The required @status specifies whether the given element is required, allowed, or disallowed. Subjects and objects may be defined according to an item type (the name of the element expected, in @item-type). An object may be defined as a datatype (@content-datatype), with an optional regular expression defining the datatype's lexical constraint (@content-lexical-constraint).

Error reporting on claims has now been generalized according to constraints specified under a &lt;verb>'s vocabulary.

Claims may now take @xml:id, allowing them to be referenced as vocabulary, and to be made parts of claims. Therefore claims may now not only nest, but become the subject or object of claims that are not ancestors.

Changes according to the points have been made above to the examples, the error tests, the schemas, and the functions.